+++ np.Templating/CHANGELOG.md
@@ -3,8 +3,9 @@
 ### About np.Templating Plugin
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/np.Templating/README.md) for details on available commands and use case.
 
-## [1.5.0-alpha.01] - 2022-06-11 (mikeerickson)
+## [1.5.0-alpha.01] - 2022-06-05 .. 2022-06-11 (mikeerickson)
 
+- added template code block execution
 - added `include` method (will include project notes, calendar notes, templates)
   > when "including" template, it will be rendered automatically
 - added `template` method